Hotfix
  - added the updated developer_readme to the project doc
  - moved some api tests from the documentation to the test suite
    also fixed them and removed a redundant one.
