+++ doc/API/authentication.md
