+++ doc/API/search.md
