+++ src/Catrobat/AppBundle/Features/Api/program_details.feature
@@ -59,4 +59,4 @@ Feature: Get details for a specific program
             "preHeaderMessages": ""
         }
       """
-      
+
